+++ src/ui/control/fullscreen_control.ts
@@ -14,6 +14,12 @@ export type FullscreenControlOptions = {
      * `container` is the [compatible DOM element](https://developer.mozilla.org/en-US/docs/Web/API/Element/requestFullScreen#Compatible_elements) which should be made full screen. By default, the map container element will be made full screen.
      */
     container?: HTMLElement;
+    /**
+     * If `true`, the fullscreen control will always use pseudo fullscreen mode (CSS-based, expanding to browser viewport) instead of native fullscreen API.
+     * This can be useful for faster transitions and to allow multiple maps to be "fullscreen" simultaneously in different browser windows.
+     * @defaultValue false
+     */
+    pseudo?: boolean;
 };
 
 /**
@@ -45,6 +51,7 @@ export class FullscreenControl extends Evented implements IControl {
     _fullscreenButton: HTMLButtonElement;
     _container: HTMLElement;
     _prevCooperativeGesturesEnabled: boolean;
+    _pseudo: boolean;
 
     /**
      * @param options - the control's options
@@ -52,6 +59,7 @@ export class FullscreenControl extends Evented implements IControl {
     constructor(options: FullscreenControlOptions = {}) {
         super();
         this._fullscreen = false;
+        this._pseudo = options.pseudo ?? false;
 
         if (options && options.container) {
             if (options.container instanceof HTMLElement) {
@@ -112,11 +120,10 @@ export class FullscreenControl extends Evented implements IControl {
     }
 
     _onFullscreenChange = () => {
+        // WebKit due to https://caniuse.com/mdn-api_document_fullscreenelement
         let fullscreenElement =
             window.document.fullscreenElement ||
-            (window.document as any).mozFullScreenElement ||
-            (window.document as any).webkitFullscreenElement ||
-            (window.document as any).msFullscreenElement;
+            (window.document as any).webkitFullscreenElement;
 
         while (fullscreenElement?.shadowRoot?.fullscreenElement) {
             fullscreenElement = fullscreenElement.shadowRoot.fullscreenElement;
@@ -154,13 +161,12 @@ export class FullscreenControl extends Evented implements IControl {
     };
 
     _exitFullscreen() {
-        if (window.document.exitFullscreen) {
+        if (this._pseudo) {
+            this._togglePseudoFullScreen();
+        } else if (window.document.exitFullscreen) {
             (window.document as any).exitFullscreen();
-        } else if ((window.document as any).mozCancelFullScreen) {
-            (window.document as any).mozCancelFullScreen();
-        } else if ((window.document as any).msExitFullscreen) {
-            (window.document as any).msExitFullscreen();
         } else if ((window.document as any).webkitCancelFullScreen) {
+            // due to https://caniuse.com/mdn-api_document_exitfullscreen
             (window.document as any).webkitCancelFullScreen();
         } else {
             this._togglePseudoFullScreen();
@@ -168,13 +174,12 @@ export class FullscreenControl extends Evented implements IControl {
     }
 
     _requestFullscreen() {
-        if (this._container.requestFullscreen) {
+        if (this._pseudo) {
+            this._togglePseudoFullScreen();
+        } else if (this._container.requestFullscreen) {
             this._container.requestFullscreen();
-        } else if ((this._container as any).mozRequestFullScreen) {
-            (this._container as any).mozRequestFullScreen();
-        } else if ((this._container as any).msRequestFullscreen) {
-            (this._container as any).msRequestFullscreen();
         } else if ((this._container as any).webkitRequestFullscreen) {
+            // due to https://caniuse.com/mdn-api_element_requestfullscreen
             (this._container as any).webkitRequestFullscreen();
         } else {
             this._togglePseudoFullScreen();